+++ fade-text-components/src/components/fade-text.js
@@ -2,6 +2,7 @@ import styled, { keyframes } from 'styled-components'
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import { screen } from '../styles/screen'
+import scrollManager from '../utils/scroll-manager'
 
 
 const FadeOut = keyframes`
@@ -29,16 +30,14 @@ const Container = styled.div`
   width: 100%;
   height: 100%;
   position: fixed;
-  z-index: 999;
+  z-index: 500;
+  top: 0;
+  left: 0;
+  padding: 0;
+  margin: 0;
   body: {
     overflow: hidden;
   }
-  ${screen.tablet`
-    height: 85%;
-  `}
-  ${screen.mobile`
-    height: 79%;
-  `}
   animation: ${Shrink} 100ms linear;
   animation-delay: ${props => (props.delay ? `${props.delay}s` : '10s')};
   animation-fill-mode: forwards;
@@ -90,7 +89,7 @@ const TextFrame = styled.div`
   text-align: center;
   width: 88%;
   font-size: 50px;
-  font-weight: 200;
+  font-weight: 100;
   line-height: 1.6;
   text-align: center;
   ${screen.desktop`
@@ -101,17 +100,60 @@ const TextFrame = styled.div`
     line-height: 1.67;
   `}
 `
+const { childAnimationStoper, unlockAfterAnimation } = scrollManager
+
+const scrollUnlocker = (e) => {
+  if (e) {
+    e.stopPropagation()
+  }
+  const body = document.body
+  const html = document.documentElement
+  body.style.overflow = 'visible'
+  body.style.height = 'auto'
+  body.style.touchAction = 'auto'
+  body.style.position = 'static'
+  html.style.overflow = 'visible'
+  html.style.height = 'auto'
+  html.style.touchAction = 'auto'
+  html.style.position = 'static'
+}
+
+const unlockScroll = (ifUnlock, node) => {
+  if (ifUnlock) {
+    unlockAfterAnimation(node, scrollUnlocker)
+  } else {
+    childAnimationStoper(node)
+  }
+}
 
 class FadeText extends PureComponent {
+  constructor(props) {
+    super(props)
+    this.global = () => {}
+  }
   componentDidMount() {
-    const elem = document.getElementById('pageContainer')
-    elem.style.overflowY = 'hidden'
-    elem.style.height = '100vh'
+    if (this.props.ifLock) {
+      const body = document.body
+      const html = document.documentElement
+      body.style.overflow = 'hidden'
+      body.style.height = '100%'
+      body.style.touchAction = 'manipulation'
+      body.style.position = 'relative'
+      html.style.overflow = 'hidden'
+      html.style.height = '100%'
+      html.style.touchAction = 'manipulation'
+      html.style.position = 'relative'
+    }
+  }
+
+  componentWillUnmount() {
+    if (this.props.ifUnlock) {
+      scrollUnlocker()
+    }
   }
 
   render() {
-    const { bgColor, duration, textArr, fontColor, bgOutDuration } = this.props
-
+    const { bgColor, duration, textArr, fontColor, bgOutDuration, ifUnlock } = this.props
     const textFrames = () => {
       return textArr.map((v, i) => {
         return (
@@ -120,6 +162,7 @@ class FadeText extends PureComponent {
             delay={i * duration}
             key={`fade_text_${v}`}
             fontColor={fontColor}
+            innerRef={childAnimationStoper}
           >
             {v}
           </TextFrame>
@@ -128,8 +171,15 @@ class FadeText extends PureComponent {
     }
 
     return (
-      <Container delay={(duration * textArr.length) + bgOutDuration + FadeOutDuration}>
-        <FadeOutContainer delay={(duration * textArr.length) + bgOutDuration}>
+      <Container
+        delay={(duration * textArr.length) + bgOutDuration + FadeOutDuration}
+        onTouchStart={(e) => { e.preventDefault() }}
+        onTouchMove={(e) => { e.preventDefault() }}
+      >
+        <FadeOutContainer
+          delay={(duration * textArr.length) + bgOutDuration}
+          innerRef={(node) => { unlockScroll(ifUnlock, node) }}
+        >
           <Background bgColor={bgColor}>
             {textFrames()}
           </Background>
@@ -145,6 +195,8 @@ FadeText.propTypes = {
   fontColor: PropTypes.string.isRequired,
   duration: PropTypes.number.isRequired,
   bgOutDuration: PropTypes.number.isRequired,
+  ifLock: PropTypes.bool.isRequired,
+  ifUnlock: PropTypes.bool.isRequired,
 }
 
 export default FadeText