+++ fade-text-components/src/index.js
@@ -1,3 +1,7 @@
 import FadeText from './components/fade-text'
+import scrollManager from './utils/scroll-manager'
 
-export default FadeText
+export default {
+  component: FadeText,
+  scrollManager,
+}